// src/shared/models/health-library.ts
export interface HealthLibrary {
    id?: any;
    image?: string;
    title?: string;
    subTitle?: string;
    date?: any;
    publishDate?: any;
    published?: boolean;
    content?: string;
    text?: string;
    space?: any;
    spaceId?: string;
    publisher?: any;
    publisherId?: string;
    isPublished?: boolean;
}